Precision logic bug.

+++ picme/Models/Singletons/FlickrAPIManager.m
@@ -76,7 +76,7 @@ static FlickrAPIManager * sharedManager_;
 
 #pragma mark - OFFlickrAPIMethods
 
-+ (void) photosForLocationWithLatitude:(NSInteger) latitude  longitude:(NSInteger) longitude resultLimit:(NSInteger) limit success:(void (^) (NSDictionary * responseDictionary)) success failure:(void (^) (NSError * error)) failure {
++ (void) photosForLocationWithLatitude:(float) latitude  longitude:(float) longitude resultLimit:(NSInteger) limit success:(void (^) (NSDictionary * responseDictionary)) success failure:(void (^) (NSError * error)) failure {
     [FlickrAPIManager sharedManager].apiSuccessBlock = success;
     [FlickrAPIManager sharedManager].apiFailureBlock = failure;
